Update vc-recaptcha.js
Recaptcha.reload requires a 't' be passed in to stop refocusing the response field. This argument is undocumented. It may have been the original authors intent to have many 'options' passed in as a string. e.g. "trnl" I changed this to be truthy but risk the api will add more options. Nothing an `===true` won't handle.