Check for grecaptcha.render function availability
Instead of checking for the variable `grecaptcha` to be defined, now we check for the `grecaptcha.render` function to be defined and to be a function.

It has happened in Firefox(due to browser cache), that sometimes the `grecaptcha` variable is already defined, but as empty object, causing the library to fail due to the lack of the render function.